added warning message for removing multiple objects

+++ simulator-ui/src/java/piccoloX/ca/shu/ui/lib/actions/RemoveObjectsAction.java
@@ -2,7 +2,10 @@ package ca.shu.ui.lib.actions;
 
 import java.util.Collection;
 
+import javax.swing.JOptionPane;
+
 import ca.shu.ui.lib.objects.models.ModelObject;
+import ca.shu.ui.lib.util.UIEnvironment;
 import ca.shu.ui.lib.world.WorldObject;
 
 public class RemoveObjectsAction extends StandardAction {
@@ -17,20 +20,26 @@ public class RemoveObjectsAction extends StandardAction {
 
 	@Override
 	protected void action() throws ActionException {
-		for (WorldObject wo : objectsToRemove) {
-			if (wo instanceof ModelObject) {
-				/*
-				 * If it's a model, make sure that's destroyed as well
-				 */
-				((ModelObject) wo).destroyModel();
-			} else {
-				/*
-				 * Just destroy the UI representation
-				 */
-				wo.destroy();
+
+		int response = JOptionPane.showConfirmDialog(UIEnvironment.getInstance(),
+				"You are about to remove " + objectsToRemove.size() + " Objects.",
+				"Continue?", JOptionPane.YES_NO_OPTION);
+
+		if (response == JOptionPane.YES_OPTION) {
+
+			for (WorldObject wo : objectsToRemove) {
+				if (wo instanceof ModelObject) {
+					/*
+					 * If it's a model, make sure that's destroyed as well
+					 */
+					((ModelObject) wo).destroyModel();
+				} else {
+					/*
+					 * Just destroy the UI representation
+					 */
+					wo.destroy();
+				}
 			}
 		}
-
 	}
-
 }
